requirements.txt
Django==1.11
djangorestframework==3.10.3
Pillow==6.2.1
psycopg2==2.8.4
pytz==2019.3
whitenoise==4.1.4
